requirements.txt
fastapi[standard]
mysql-connector-python
PyJwt
aiocache